fix: close hosted validation gaps
